feat(metrics): Add better error metrics
Previously our error metrics only exposed three kinds (Internal, Auth,

and Invalid request) and thus wasn't super insightful.
This commit adds a new pattern: `*ErrorMetric` types that don't contain
inner dynamic data, ie, are essentially just static strings with low
cardinality. This makes it possible to put an upper bound on the
cardinality of the error types our metrics receive which is required to
not cause prometheus to slow to a crawl.

From impls are added from `&OriginalErrorType` to the `*ErrorMetric`
type to make conversion cheap and easy and is combined with a top level
`ErrorMetric` trait to get the final error string to report the error
type/kind in Opentelemetry.