feat: restructure gw docs
the intention with this change is to emphasize the new cloud gateway and
de emphasize the self hosted gateway.

one thing i didnt know how to achieve was make it so the sidebar group
for self hosted deployment (a nested group) could have the root page be
`/ai-gateway/self-hosted/deployment/overview`. from what i can see in
the mintlify docs, this isn't possible, but i think it would be
preferable so that if you click "deployment options" it brings you to
the overview while *also* expanding the group, AND without taking up two
lines in the sidebar