fix: Update WITH FILL clause to use correct data types
Currently, these queries using the WITH FILL clause do not error on our
current version of Clickhouse (v24), presumably, clickhouse is doing
some casting for us.

However, in the next version of Clickhouse (v25), instead of casting for
us, Clickhouse returns an error, meaning we need to update our query and
do the transformation/casting on our end, which is achieved in this
commit.

this is tested against both v24 and v25 of clickhouse locally and
confirmed to work