fix: map status type depending on anthropic error
unfortunately, the anthropic streaming api will always return a 200
status code even when the actual response body is an error. in those
cases, the response body looks like this, for example:

```
{
  "streamed_data": "event: error\ndata: {\"type\":\"error\",\"error\":{\"details\":null,\"type\":\"overloaded_error\",\"message\":\"Overloaded\"}              }\n\n"
}
```

this commit updates the llm mapper to correctly determine the
`statusType` is an error even when the status code is a 200.

this ofc does come at the risk of marking a response as an error if a
valid, successful llm response contains the literal `overloaded_error`