fix: Override status for errors in anthropic streams
context: anthropic will always return a 200 status code, even when there
is actually an error sent in the event stream itself. this would confuse
customers as when looking at the logs in the Helicone dashboard,
they would expect to see a response since they see a 200 status code,
but instead would see some random JSON (the error). to fix this, we are
allowing the stream body parsers to override the status code.